+++ controllers/routine_controller.py
@@ -1,8 +1,10 @@
 from flask import Blueprint, request
 from flask_jwt_extended import jwt_required, get_jwt_identity
+from marshmallow.exceptions import ValidationError
 
 from init import db
-from models.routine import Routine, routines_schema, routine_schema
+from models.routine import Routine, routines_schema, routine_schema, WEEKDAYS
+from models.user import User
 
 from controllers.exercise_controller import exercise_bp
 
@@ -24,7 +26,7 @@ def get_routine_byID(routine_id):
     if one_routine:
         return routine_schema.dump(one_routine)
     else:
-        return {"error": f"Routine with '{routine_id}' hasn't been setup yet"}, 404
+        return {"error": f"Routine with 'id {routine_id}' hasn't been setup yet"}, 404
     
 @routine_bp.route("/name/<routine_name>")
 def get_routine_byName(routine_name):
@@ -49,31 +51,60 @@ def get_routine_byDay(routine_Day):
 @routine_bp.route("/", methods=["POST"])
 @jwt_required()
 def create_new_routine():
-    body_data = request.get_json()
-    # body_data = routine_schema.load(request.get_json())
-    # Create a new card model instance
+    body_data = routine_schema.load(request.get_json())
+    stmt = db.select(Routine)
+    routine = db.session.scalar(stmt)
+    
+    # Create a new routine model instance
     routine = Routine(
-        name = body_data.get("name"), # it's unique have to handle errors
+        name = body_data.get("name"),
         description = body_data.get("description"),
         weekday = body_data.get("weekday"),
         user_id = get_jwt_identity()
     )
+    
+    # name_stmt = db.select(db.func.count()).select_from(Routine).filter_by(name=routine.name)
+    # # day_stmt = db.select(db.func.count()).select_from(Routine).filter_by(weekday=routine.weekday)
+    # repeated_routine_name = db.session.scalar(name_stmt)
+    # # repeated_routine_day = db.session.scalar(day_stmt)
+    
+    # if repeated_routine_name >= 1:
+    # if str(routine.user_id) == get_jwt_identity():
+    #     print(routine.user_id)
+    #     raise ValidationError(f"'{routine.name}' routine for {routine.user_id} already.")
+        # elif repeated_routine_day > 0:
+        #     raise ValidationError(f"There's a routine for '{routine.weekday}'. Only one routine per day")
+    
+    
+    
     # add to the session and commit
     db.session.add(routine)
     db.session.commit()
-    # return the newly cerated card
+    # return the newly cerated routine
     return routine_schema.dump(routine), 201
         
 # The Update - part of CRUD
 @routine_bp.route("/<int:routine_id>", methods=["PATCH", "PUT"])
 @jwt_required()
 def update_routine(routine_id):
-    body_data = request.get_json()
+    body_data = routine_schema.load(request.get_json(), partial=True)
     
     stmt = db.select(Routine).filter_by(id = routine_id)
     routine = db.session.scalar(stmt)
     
     if routine:
+        if str(routine.user_id) != get_jwt_identity():
+            print(routine.user)
+            return {"error": "Only the creator of the routine can edit it"}, 403
+        
+        # if routine.weekday in WEEKDAYS:
+        #     index_of_routine_day_in_WEEKDAYS = WEEKDAYS.index(routine.weekday)
+        #     stmt = db.select(db.func.count()).select_from(Routine).filter_by(weekday=WEEKDAYS[index_of_routine_day_in_WEEKDAYS])
+        #     routines_in_day_count = db.session.scalar(stmt)
+            
+        #     if routines_in_day_count > 0:
+        #         raise ValidationError(f"There's a routine for {routine.weekday} already. Only one routine per day")
+    
         routine.name = body_data.get("name") or routine.name
         routine.description = body_data.get("description") or routine.description
         routine.weekday = body_data.get("weekday") or routine.weekday
@@ -90,8 +121,11 @@ def delete_routine(routine_id):
     stmt = db.select(Routine).where(Routine.id == routine_id)
     routine = db.session.scalar(stmt)
     if routine:
+        if str(routine.user_id) != get_jwt_identity():
+            return {"error": "Only the creator can delete this routine"}, 403
         db.session.delete(routine)
         db.session.commit() 
         return {"message": f"{routine.name} routine for {routine.weekday} has now been deleted"}
     else:
         return {"message": f"Routine not found"}, 404
+
