+++ controllers/sets_reps_controller.py
@@ -4,6 +4,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from init import db, bcrypt
 from models.exercise import Exercise
 from models.sets_reps import SetsReps, set_rep_schema
+from models.user import User
 
 
 sets_reps_bp = Blueprint("set_and_reps", __name__, url_prefix="/<int:exercise_id>/sets_and_reps")
@@ -19,7 +20,7 @@ sets_reps_bp = Blueprint("set_and_reps", __name__, url_prefix="/<int:exercise_id
 @sets_reps_bp.route("/", methods=["POST"])
 @jwt_required()
 def create_sets_reps(exercise_id):
-    body_data = request.get_json()
+    body_data = set_rep_schema.load(request.get_json(), partial=True)
     stmt = db.select(Exercise).filter_by(id=exercise_id)
     exercise = db.session.scalar(stmt)
     
@@ -27,7 +28,8 @@ def create_sets_reps(exercise_id):
     set_rep = db.session.scalar(stmt_set_rep)
     
     if exercise.sets_reps:
-        return [{"error": f"There are asigned sets and repetitions to {exercise.name}"},{"Sets and Reps Assigned": f"set: {set_rep.sets}, repetitons: {set_rep.reps}, goal: {set_rep.goal}"}]
+        return [{"error": f"There are asigned sets and reps to '{exercise.name}'"},{"Sets and Reps Assigned": f"set: {set_rep.sets}, repetitons: {set_rep.reps}, goal: {set_rep.goal}"}], 403
+    
     if exercise:
         set_rep = SetsReps(
             sets = body_data.get("sets"),
@@ -40,7 +42,7 @@ def create_sets_reps(exercise_id):
         db.session.commit()
         return set_rep_schema.dump(set_rep), 201
     else:
-        return {"error": f"{exercise.name} not found"}, 404
+        return {"error": f"'{exercise.name}' not found"}, 404
 
 @sets_reps_bp.route("/", methods=["DELETE"])
 @jwt_required()
@@ -51,21 +53,27 @@ def delete_sets_reps(exercise_id):
     stmt_exercise = db.select(Exercise).filter_by(id=exercise_id)
     exercise = db.session.scalar(stmt_exercise)
     
-    if set_rep:
+    if not exercise:
+        return {"error": f"Not found, sets or reps are not yet set up"}, 404
+    elif exercise and set_rep:
+        if str(exercise.user_id) != get_jwt_identity():
+            return {"error": f"Only the creator of '{exercise.name}' can delete its associted sets and reps"}, 403
         db.session.delete(set_rep)
         db.session.commit()
-        return {"message": f"sets and reps for {exercise.name} have been deleted"}
+        return {"message": f"sets and reps for '{exercise.name}' have been deleted"}
     else:
-        return {"error": f"sets and repetitions are not yet set up for {exercise.name}"}, 404
+        return {"error": f"Not found, sets or reps are not yet set up for this exercise"}, 404
     
 @sets_reps_bp.route("/<int:sets_reps_id>", methods=["PATCH", "PUT"])
 @jwt_required()
 def edit_sets_reps(exercise_id, sets_reps_id):
-    body_data = request.get_json()
+    body_data = set_rep_schema.load(request.get_json(), partial=True)
     stmt = db.select(SetsReps).filter_by(id=sets_reps_id, exercise_id=exercise_id)
     set_rep = db.session.scalar(stmt)
     
     if set_rep:
+        if str(set_rep.user_id) != get_jwt_identity():
+            return {"error": f"These sets and reps are associted to another user's exercise, you cannot modify it"}, 403
         set_rep.sets = body_data.get("sets") or set_rep.sets
         set_rep.reps = body_data.get("reps") or set_rep.reps
         set_rep.goal = body_data.get("goal") or set_rep.goal
@@ -73,4 +81,4 @@ def edit_sets_reps(exercise_id, sets_reps_id):
         db.session.commit()
         return set_rep_schema.dump(set_rep)
     else:
-        return {"error": f"No sets and repetitions found in '{exercise_id.name}'"}, 404
+        return {"error": f"No sets and reps found"}, 404
